+++ lib/screens/settings/settings_screen.dart
@@ -89,7 +89,7 @@ class _SettingsScreenState extends State<SettingsScreen> {
                 SettingsTile.navigation(
                   // leading: const Icon(Icons.person_outlined),
                   title: Text(S.of(context).inventories),
-                  description: Text('${S.current.limits} • ${S.current.defaults} • ${S.current.reminders}'),
+                  description: Text('${S.current.inventoryRules} • ${S.current.defaults} • ${S.current.reminders}'),
                   onPressed: (context) {
                     Navigator.push(
                       context,