+++ assets/changelog.md
@@ -15,6 +15,7 @@
 * Added in settings an option to define what to do with existing records when importing data (ask, update or ignore).
 * Added tags for field journal notes.
 * Added option in settings to edit the available tags for field journal notes.
+* Added in settings an option to define the policy for species propagation from inactive inventories (ask, propagate or ignore).
 
 ### Improvements